simplify CtaBar and HeaderDueDateBar
